# final/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Productos",
        "description": "Listado de endpoints de los productos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listado de todos los registros de productos",
                "operationId": "ed2ea672fb88685f1a9ed0b401f15fe5",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "nombres": {
                                                        "type": "string",
                                                        "example": "Aderson Felix"
                                                    },
                                                    "descripcion": {
                                                        "type": "string",
                                                        "example": "Jara Lazaro"
                                                    },
                                                    "precio_unitario": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "cantidad": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "costo_total": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T00:09:16.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-02-23T12:33:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un nuevo producto con la información proporcionada.",
                "operationId": "358072a9b20fa5b5c9c1c063e4fb0469",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "precio_unitario": {
                                        "type": "number"
                                    },
                                    "cantidad": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto creado con éxito"
                                        },
                                        "product": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "nombre": {
                                                    "type": "string"
                                                },
                                                "descripcion": {
                                                    "type": "string"
                                                },
                                                "precio_unitario": {
                                                    "type": "number"
                                                },
                                                "cantidad": {
                                                    "type": "integer"
                                                },
                                                "costo_total": {
                                                    "type": "number"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error de validación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/productos/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Mostrar la información de un cliente",
                "operationId": "b657e1b29ce26261bb31d11477da49bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nombre": {
                                            "type": "string",
                                            "example": "Aderson Felix"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Jara Lazaro"
                                        },
                                        "precio_unitario": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "cantidad": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "costo_total": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "PRODUCTO NO ENCONTRADO",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto existente",
                "description": "Actualiza un producto existente con la información proporcionada.",
                "operationId": "102f75959a9b57248cd8e60e825b6215",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "precio_unitario": {
                                        "type": "number"
                                    },
                                    "cantidad": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto actualizado con éxito"
                                        },
                                        "product": {
                                            "properties": {
                                                "id": {
                                                    "type": "number"
                                                },
                                                "nombre": {
                                                    "type": "string"
                                                },
                                                "descripcion": {
                                                    "type": "string"
                                                },
                                                "precio_unitario": {
                                                    "type": "number"
                                                },
                                                "cantidad": {
                                                    "type": "integer"
                                                },
                                                "costo_total": {
                                                    "type": "number"
                                                },
                                                "created_at": {
                                                    "type": "string"
                                                },
                                                "updated_at": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Mensaje de error de validación"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "PRODUCTO NO ENCONTRADO",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Producto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto",
                "operationId": "df607d9b875d022ccd28f3939c7cb2e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto eliminado con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "PRODUCTO NO ENCONTRADO",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}